+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.7;
-const unsigned char AutoSDKVersionString[] = "1.7.0";
+double AutoSDKVersionNumber = 1.8;
+const unsigned char AutoSDKVersionString[] = "1.8.0";