+++ Sources/AutoSDK/AutoScriptSupport.h
@@ -16,6 +16,8 @@ NSString * _Nullable AutoScriptMD5Hex(NSData *data);
 NSString * _Nullable AutoScriptSHA1Hex(NSData *data);
 NSString * _Nullable AutoScriptSHA256Hex(NSData *data);
 NSString * _Nullable AutoScriptSHA512Hex(NSData *data);
+NSString * _Nullable AutoScriptHMACSHA1Hex(NSString *text, NSString *key);
+NSString * _Nullable AutoScriptHMACSHA256Hex(NSString *text, NSString *key);
 NSString * _Nullable AutoScriptAES128EncryptBase64(NSString *plaintext, NSString *key);
 NSString * _Nullable AutoScriptAES128DecryptBase64(NSString *base64, NSString *key);
 NSString * _Nullable AutoScriptToPinYin(NSString *text);